+++ src/conn_core/connect_params.rs
@@ -4,7 +4,6 @@ use crate::{HdbErrorKind, HdbResult};
 use failure::ResultExt;
 use secstr::SecStr;
 use std::env;
-use std::fmt;
 use std::fs;
 use std::path::Path;
 use url::Url;
@@ -20,8 +19,9 @@ use url::Url;
 /// ### Example with TLS
 ///
 /// ```rust,ignore
+/// # // norun cannot be used because the snippet requires feature tls
 /// use hdbconnect::{ConnectParams, ServerCerts};
-/// # fn read_certificate() -> String {unimplemented!()};
+/// # fn read_certificate() -> String {String::from("can't do that")};
 /// let certificate: String = read_certificate();
 /// let connect_params = ConnectParams::builder()
 ///    .hostname("the_host")
@@ -77,7 +77,7 @@ use url::Url;
 /// The shortcut [`ConnectParams::from_file`](struct.ConnectParams.html#method.from_file)
 /// reads a URL from a file and converts it into an instance of `ConnectParams`.
 ///
-#[derive(Clone)]
+#[derive(Clone, Debug)]
 pub struct ConnectParams {
     host: String,
     addr: String,
@@ -176,8 +176,8 @@ impl ConnectParams {
     }
 }
 
-impl fmt::Debug for ConnectParams {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+impl std::fmt::Display for ConnectParams {
+    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
         write!(
             f,
             "ConnectParams {{ addr: {}, dbuser: {}, clientlocale: {:?} }}",