+++ src/types_impl/lob/lob_writer.rs
@@ -221,11 +221,9 @@ impl<'a> Write for LobWriter<'a> {
                 payload_raw
             };
 
-            let locator_ids = self
-                .write_a_lob_chunk(LobWriteMode::Append(&payload))
+            self.write_a_lob_chunk(LobWriteMode::Append(&payload))
+                .map(|_locator_ids| ())
                 .map_err(|e| util::io_error(e.to_string()))?;
-            debug_assert_eq!(locator_ids.len(), 1);
-            debug_assert_eq!(locator_ids[0], self.locator_id);
         }
         Ok(input.len())
     }
@@ -248,10 +246,9 @@ impl<'a> Write for LobWriter<'a> {
             payload_raw
         };
 
-        let locator_ids = self
-            .write_a_lob_chunk(LobWriteMode::Last(&payload))
+        self.write_a_lob_chunk(LobWriteMode::Last(&payload))
+            .map(|_locator_ids| ())
             .map_err(|e| std::io::Error::new(std::io::ErrorKind::Other, e.to_string()))?;
-        debug_assert_eq!(locator_ids.len(), 0);
         Ok(())
     }
 }