+++ tests/test_034_nclobs.rs
@@ -14,6 +14,7 @@ use std::io::Read;
 #[test]
 pub fn test_034_nclobs() -> HdbResult<()> {
     let mut log_handle = test_utils::init_logger();
+    let start = std::time::Instant::now();
     let mut connection = test_utils::get_authenticated_connection()?;
 
     let (blabla, fingerprint) = get_blabla();
@@ -26,8 +27,7 @@ pub fn test_034_nclobs() -> HdbResult<()> {
     )?;
     test_bytes_to_nclobs(&mut log_handle, &mut connection)?;
 
-    info!("{} calls to DB were executed", connection.get_call_count()?);
-    Ok(())
+    test_utils::closing_info(connection, start)
 }
 
 fn get_blabla() -> (String, Vec<u8>) {