+++ tests/test_070_explain.rs
@@ -7,11 +7,11 @@ use log::{debug, info};
 #[test]
 pub fn test_070_explain() -> HdbResult<()> {
     let mut log_handle = test_utils::init_logger();
+    let start = std::time::Instant::now();
     let mut connection = test_utils::get_authenticated_connection()?;
 
     test_explain(&mut log_handle, &mut connection)?;
-    info!("{} calls to DB were executed", connection.get_call_count()?);
-    Ok(())
+    test_utils::closing_info(connection, start)
 }
 
 fn test_explain(