+++ CHANGELOG.md
@@ -2,12 +2,15 @@
 
 ## [0.19.0-unpublished]
 
+Remove wrong debug asserts from lob_writer (adding to the fix for issue 23).
+
 Reconstruct HdbError, based on crate `failure`.
 
 Improve ConnectParamsBuilder
 
 - move root-certificates option as an additional variant into ServerCerts enum
 - don't erase any information from them builder during build()
+- implement Display, and derive Debug, for ConnectParams
 
 Minor docu improvements.